Change ID type

+++ baggage/src/main/java/com/brownfield/pss/baggage/component/BaggageComponent.java
@@ -38,7 +38,7 @@ public class BaggageComponent {
 		return 0;
 	}
 
-	public List<BaggageRecord> getBaggageOfCheckin(long checkinId) {
+	public List<BaggageRecord> getBaggageOfCheckin(int checkinId) {
 		return baggageRepository.findByCheckinId(checkinId);
 	}
